[FEAT/#23] Sheet 영역 이외를 선택해도 Sheet가 내려가지 않도록 처리
