+++ BiRTH/BiRTH/View/CollageByMyselfView/ColorManager.swift
