+++ BiRTH/BiRTH/View/CollageByMyselfView/Components/XmarkButton.swift
