+++ BiRTH/BiRTH/View/CollageByMyselfView/PastedImage.swift
