+++ BiRTH/BiRTH/View/CollageByMyselfView/TempFriendDetailView.swift
